replaced parse_latex.py with NRPyLaTeX package
