+++ requirements.txt
@@ -1,2 +1,3 @@
 sympy
 matplotlib
+nrpylatex>=1.0.2
